requirements.txt
langchain_groq
streamlit
langchain
python-docx